requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile pyproject.toml
#
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==22.2.0
    # via
    #   outcome
    #   trio
certifi==2022.12.7
    # via
    #   requests
    #   selenium
cffi==1.15.1
    # via trio
charset-normalizer==3.0.1
    # via requests
exceptiongroup==1.1.0
    # via trio
h11==0.14.0
    # via wsproto
idna==3.4
    # via
    #   requests
    #   trio
outcome==1.2.0
    # via trio
pycparser==2.21
    # via cffi
pysocks==1.7.1
    # via urllib3
requests==2.28.2
    # via robotframework-beholder (pyproject.toml)
robotframework==6.0.2
    # via
    #   robotframework-beholder (pyproject.toml)
    #   robotframework-seleniumlibrary
robotframework-pythonlibcore==4.1.2
    # via robotframework-seleniumlibrary
robotframework-seleniumlibrary==6.0.0
    # via robotframework-beholder (pyproject.toml)
selenium==4.8.2
    # via robotframework-seleniumlibrary
sniffio==1.3.0
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.22.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[socks]==1.26.14
    # via
    #   requests
    #   selenium
wsproto==1.2.0
    # via trio-websocket
